Disable core/oUF.lua for Midnight compatibility
Comments out the core/oUF.lua file in Framework.xml to prevent errors
when the UnitFrame system is disabled. This file contains custom oUF
elements and tags that depend on the SUIUF global being defined, which
is only available when the oUF library is loaded.

With oUF disabled for Midnight builds:
- libs/oUF/oUF.xml is commented out in SpartanUI.toc
- Modules/UnitFrames is commented out in Modules/LoadAll.xml
- Modules/Nameplates.lua is commented out in Modules/LoadAll.xml
- core/oUF.lua is now commented out in Core/Framework.xml

This prevents "attempt to index global 'SUIUF' (a nil value)" errors.